% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_fixtures.R
\name{.hzr_create_lognormal_golden_fixture}
\alias{.hzr_create_lognormal_golden_fixture}
\title{Create and save log-normal golden fixture}
\usage{
.hzr_create_lognormal_golden_fixture(output_dir = NULL)
}
\description{
Create and save log-normal golden fixture
}
\keyword{internal}
