% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase-spec.R
\name{is_hzr_phase}
\alias{is_hzr_phase}
\title{Test if an object is an hzr_phase}
\usage{
is_hzr_phase(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test if an object is an hzr_phase
}
\examples{
is_hzr_phase(hzr_phase("cdf"))
is_hzr_phase("not a phase")

}
