% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_api.R
\name{coef.hazard}
\alias{coef.hazard}
\title{Extract coefficients from hazard model}
\usage{
\method{coef}{hazard}(object, ...)
}
\arguments{
\item{object}{A \code{hazard} object.}

\item{...}{Unused; for S3 compatibility.}
}
\description{
Extract coefficients from hazard model
}
