% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_primitives.R
\name{hzr_log1mexp}
\alias{hzr_log1mexp}
\title{Numerically stable log(1 - exp(-x)) for x > 0}
\usage{
hzr_log1mexp(x)
}
\arguments{
\item{x}{Numeric vector with positive values.}
}
\value{
Numeric vector with element-wise \code{log(1 - exp(-x))}.
}
\description{
Numerically stable log(1 - exp(-x)) for x > 0
}
