% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_primitives.R
\name{hzr_log1pexp}
\alias{hzr_log1pexp}
\title{Numerically stable log(1 + exp(x))}
\usage{
hzr_log1pexp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with element-wise \code{log(1 + exp(x))}.
}
\description{
Numerically stable log(1 + exp(x))
}
