% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.hzr_calibrate}
\alias{print.hzr_calibrate}
\title{Print method for hzr_calibrate}
\usage{
\method{print}{hzr_calibrate}(x, digits = 3, ...)
}
\arguments{
\item{x}{An \code{hzr_calibrate} object.}

\item{digits}{Number of significant digits.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for hzr_calibrate
}
