% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.hzr_deciles}
\alias{print.hzr_deciles}
\title{Print method for hzr_deciles}
\usage{
\method{print}{hzr_deciles}(x, digits = 3, ...)
}
\arguments{
\item{x}{An \code{hzr_deciles} object.}

\item{digits}{Number of decimal places for formatting.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for hzr_deciles
}
