% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.hzr_gof}
\alias{print.hzr_gof}
\title{Print method for hzr_gof}
\usage{
\method{print}{hzr_gof}(x, digits = 3, ...)
}
\arguments{
\item{x}{An \code{hzr_gof} object.}

\item{digits}{Number of significant digits for formatting.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for hzr_gof
}
