% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.hzr_kaplan}
\alias{print.hzr_kaplan}
\title{Print method for hzr_kaplan}
\usage{
\method{print}{hzr_kaplan}(x, digits = 4, n = 20, ...)
}
\arguments{
\item{x}{An \code{hzr_kaplan} object.}

\item{digits}{Number of significant digits.}

\item{n}{Maximum rows to print (default 20).}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for hzr_kaplan
}
